loader: try different allocation addresses.
This fixes #16 by repeating different memory allocations until
it finds a contiguous 512 MB chunk, and then builds the kernel
command line to add memmap=exactmap for this address.

The chainload tool now takes an argument for where to locate the
entry point and the UEFI context is extended to include the pointer
to this base.